app_python/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp

